Import everything else from the anaconda kickstart_tests directory.
I'm doing this to have a point in the repo's history where everything is
exactly as it was in anaconda.  I haven't run it to see if it works with
this repo layout.  I expect it doesn't.  After this commit, the history
will reflect whatever I had to do to make it all work.