Add --fsoptions to a couple lvm, part, and raid commands.
We only need to test this once for each command, to make sure it gets
passed through to the fstab file.  I've added these to tests that were
already doing custom partitioning instead of adding a whole new test.

Also note that I haven't added it to a btrfs test - the option isn't
supported, which may be a bug.