Don't return a failure code on kernel panics.
They happen with some frequency, and they're obviously kernel bugs and
nothing we can deal with.  Sometimes kernel panics aren't even fatal,
but we can't tell the difference between ones that are and ones that are
not.  Thus kernel panics shouldn't keep our tests from passing.  If
something else is going wrong that prevents the test from finishing,
we should still see that.