proxy: log host and port for CONNECT requests
TLS proxying is done with CONNECT requests, where all the proxy
does is open a tunnel to the remote server and let the client
and the server negotiate a TLS connection. The simple proxy for
the proxy tests implements this, but doesn't log it at all, so
we can't test that TLS requests were proxied.

Since this is, after all, TLS traffic, there's no way we can
log the actual request locations, but we can at least log the
host and port whenever a CONNECT request is handled, so let's
do that much.