rework proxy test generation, make tests more robust
While working on running kickstart-tests in openQA we noticed
that the proxy tests had some unspoken assumptions: that the
repo URL would be a direct mirror URL (not a mirror list) and
that it would use HTTP (not HTTPS).

While fixing that, I noticed there's quite a bit of duplication
between proxy-auth and proxy-kickstart, and proxy-cmdline is
more or less a subset of them only they had diverged quite a
lot in style. I also thought the mechanism for getting the proxy
code into the kickstart %pre - which requires both make *and*
ksflatten - seemed a bit over-complex for really just sticking
chunks of text together.

This is my suggestion for resolving most of the problems. It uses
a Makefile to produce all three proxy .ks.in files from the
same template plus a few snippet files and proxy.py, and makes
the tests more robust against use of mirror lists and HTTPS
for the base repo. Caveats:

* It does not cover use of HTTPS for the addon repo, that is
  still required to be served as HTTP. We *could* allow for
  this by just weakening a couple of the proxy-repo.inc tests
  for the 'httpslist' and 'httpsdir' cases, I'm not sure if
  it's worth it.

* proxy.py seems not to be capable of proxying HTTPS traffic
  with password authentication enabled, so you still can't
  do proxy-auth with an HTTPS mirror or mirrorlist. This is
  a pre-existing issue, I believe.