reqpart, default-fstype: ensure just one result from blkid
running reqpart in openQA, I found blkid would actually return
three results for this search, so the output would be 'xfs\n
xfs\nxfs' (not ext4 because I'm testing with a Server image).
The -l option protects against this by returning only one
result for the search, the one with the 'highest priority' (per
the man page).