proxy tests: handle HTTPS and mirrorlists for KSTEST_URL, unify format (#16)
* bring proxy-cmdline.ks.in in line with the other two

all three proxy tests do this same basic stuff, but the style
in cmdline was different from the other two.

* proxy tests: handle HTTPS and mirrorlist KSTEST_URL

while working on running kickstart-tests in openQA, I noticed
that they do not work if KSTEST_URL is anything but an HTTP
mirror. If it's a mirror list and/or uses HTTPS, the tests
fail. This should work as well as possible in all four cases,
and warn that complete checking is not possible for a direct
HTTPS mirror URL when that is used.