Fix driverdisk-disk test
Driver disk .ko file was generated to folder
/usr/lib/modules/3.0.0/extra/ but test pointed to location
/lib/modules/`uname -r`/updates/.

Fix location in test.

Explanation:
Module version folder is created by the RPM with driver disk. The RPM
is created by the script in lib/mkdud.py. The kernel version is 3.0.0
if version parameter is not explicitly set. Because the tests are run
on a boot.iso, we don't know which kernel version is there in time of
the script execution.
TODO: This can be improved by grabbing logs from Lorax when building
this iso or by running the boot.iso in some pre-testing phase.