Add support for multiple network devices.
It should be done by defining nics to be created by arguments of virt-install
--network options in prepare_network function. By default the function doesn't
define any nics which means no --network arguments are passed to virt-install
(as it used to be). In this case virt-install will use bridge if it finds any
or "default" network for the nic.

Future improvements:
- testing static configuration
  - learn the subnet prefix of default network in some pre stage
  - define per-test network to prevent static assignment collisions of parallell
    tests

Gotchas:
- device names: we are assuming ens3, ens4, ens5, ...