Move probe boot.iso code to separate script
Remove code for grabbing information from stage 2 on boot.iso from the
run_kickstart_tests.sh main script and move it to the separate
probe_boot_iso.sh script.
This change is required for running the main script as the root user.
Inside of the probing code there are mount and umount commands which
are required to have the root privilege.