Modify proxy-kickstart test to use squid instead of proxy.py
Some checks were moved to validate in shell script instead from the
post section. This was required because the proxy log is not inside of
tested VM machine anymore.

Added functions-proxy.sh to share checks for the proxy tests between
them.

Also after this change the proxy-kickstart test can be used on RHEL7
by resolving if it should use dnf or yum automatically.