network: add test for onboot policy
In rhel we make sure at least one the devices activated during installations
has ONBOOT set to yes.

This test handles the case where ONBOOT was set to yes for additional device
because we were looking only on ONBOOT values in ksdata and a device activated
via boot options without ks configuration is not in ksdata for automatic
installations (ie if network spoke is not visited) so we were wrongly assuming
there is no device with ONBOOT=yes.

Note: the originally failing test for this issue was network-static, but
network-static is now replaced by more comprehensive network-static-2 tests and
this issue will be tested by this new test.