Test the install method hmc with a fake support for SE/HMC
This install method can be used only on s390x with SE/HMC. To be
able to test it, we use fake scripts that will mount DVD the usual
way.

This test requires a full DVD ISO, so it is marked as a known failure.
It can be used only for accessing the packages on the DVD. It cannot
be used to get stage2.

Related: rhbz#1289918