Add initial geolocation kickstart test
This kickstart test is very simple and basically just checks
that the geolocation lookup is skipped during a kickstart installation.

Once the support for enabling geolocation during kickstart installation
is backported to Fedora we can add more tests that test the opposite,
eq. that the geolocation lookup happens & influences the installation
environment.