Always add a virtual RNG device
From 4.16 onwards the Linux kernel apparently started to be much
more strict about the amount of entropy being available which in our
case (freshly generated VM with no network traffic and no user
interaction) generally manifests as a hang at boot.

So tell virt-install to always include a virtual RNG when creating
the VMs, which should hopefully fix the issue.