Add optional log level setup for ProcessLauncher
This is required because Lorax has StreamHandler set-up to INFO log
level so the only way how to stop printing to stderr is to log to DEBUG
log.
This can't be changed outside of the library. Based on the Python 3
documentation the only handler for a library should be NullHandler.