Log to our own log not to livemedia-log
The livemedia-log prints to stderr by default and that can't be changed
easily. By logging into our own log we will avoid most of the logging
messages.

Another problem is that output from parallel is first stdout (print) and
then stderr (log) so it makes hard to read output from the tests.