Use rsync for remote copy & keep last used image around
Use rsync instead of scp to copy the kickstart-tests folder and boot iso
to the remote test runners and also keep the last used boot iso
around (we only keep the last used image and delete any others).

This way we will get a much faster test startup on remote runners as
only changes in the kickstart test will be transferred and the image
transfer skipped altogether if the same image is used.

For synchronizing the kickstart-tests folder the rsync --delete flag
if used, so no artifacts created during the tests runs in the folder
should remain once rsync is done with the sync.