Adjust tests to module availability changes
The 8 stream for the nodejs module was apparently dropped and the
nodejs no longer has two streams, which is something we need for
testing.

So switch to the avocado modules in such cases, which (as only module
in Rawhide at the moment) has two streams (stable & latest) available.

Some other modules & streams have changed or been dropped since the tests have
been initially created so adjust them accordingly.

There are also few modules that are apparently broken at the moment
in Rawhide and had to be disabled in the tests for now:

- postgres
- mariadb
- docker

Also we can't currently verify of profiles are correctly installed
for the cri-o module due to a bug in how DNF formats its output.