Add host id for kickstart test result line
When many kickstart tests are running in parallel on different
machines/VM hosts it can be very useful to know on which host
a given test was run.

This information can be used for example to debug host specific issues
that break tests (machine is too slow or unstable, has broken
networking, etc.) or to directly access test logs on machines during a
long run.

So add a host id (at the moment simply the machine hostname) to
the RESULT line, so we can easily see on which machine did the test run.

Also adjust the run_report.sh script to be compatible with the new RESULT format.