Disable some of the tests for LUKS
It looks like libvirt is not always able to work with encrypted storage
and it fails with "no operating system was found on this disk" when
it tries to copy out files. It means that the results of these tests
have to be checked manualy, so let's disable them for now.