Add script for appending of %ksappend fragments
Add the apply-ksappend.py script that uses pykickstart
to processes kickstart test files and replaces
all %ksappend lines with the kickstart fragments they
point to.

This makes it possible to both de-duplicate a lot of boilerplate
in the tests and also make support for running the tests on more
platforms easier with dynamic platform based fragments.

The script can process one or more tests, specified via the
-t/--test-file-paths option as a space separated list.

Platform (currently just fedora_rawhide) needs to be specified as
well via the -p/--platform-name option.

It is also possible to specify one or more folders with overrides
via the -o/--override-folders option. Overrides work by putting
content of the override folder on top of what is already in the
runtime folder (generally shared & platform specific fragments).

Multiple override folders can be specified as a space separated
list and are applied in order of definition:.