Use apply-ksappend.py in the main run_kickstart_tests.sh script
Run the ksappend substitution on all files after we do the other transformations
that turn a foot_test.ks.in file into the final foo_test.ks file. We
need to apply the %ksapend resolution on *.ks files as we can't rewrite
the *.ks.in files which are tracked in git.

The main script now also a new -p option option that can be used to set
platform name for resolving platform specific fragments and a -o option
for passing a white space separated list of ksappend override folders (if any).