Add initial kickstart fragments for %ksappend use
The basic idea is that all the kickstart fragments stored in the
"fragments/shared" subfolder are generally useful for running tests
across different platforms.

Content of "fragments/shared" is used to populate the empty initial runtime
directory in "fragments/runtime".

The fragments in the "fragments/platform/<platform name>" directory contain
commands, such as repository configuration, that are platform specific
and need to be substituted in kickstart test files based on the
platform the test are going to run on.

This is achieved by setting as platform name, which will result
in putting content from the corresponding "fragments/platform/<platform name>"
folder on top of all the sahred fragments already in
"fragments/runtime".

For example we want to make the repository definition platform
independent, by replacing the repo/url lines in the kickstart test file
by the following ksappend fragment substitution:

%ksappend repos/default.ks

Then when apply-ksappend.py is called with "fedora_rawhide" as platform
name, this fragment will come from:

fragments/platform/fedora_rawhide/repos/default.ks

As it will be placed to ther runtime directory when the content of
"fragments/platform/fedora_rawhide" is copied in.

The same can be easily used for other platforms such as RHEL/CentOS/etc.
in the future.