Make it possible to skip multiple test types
Till now it was possible to specify a single test type to run
and also skip a single test type.

This might not be enough if one for example wants to run
all the packaging related tests but also skip all tests marked
as both knowfailure and rhel-only.

So make it possible to pass multiple test types to skip via the
run_kickstart_tests.sh script -s option.

Please note that due to specifics of the shell option parser
multiple test types need to be passed as a quoted space delimited
string:

-s "rhel-only knownfailure"