Make sure sane locale is used on remote hosts
We use GNU Parallel to run the kickstart test jobs both locally
and on remote hosts. When it runs the jobs Parallel apparently
imports shell variables from the local host starting the job batch
to the shell environment on the remote hosts running the jobs.

This can cause issues if there is a locale availability mismatch
between the local and remote hosts, such as local host having
locale cs_CZ.UTF-8 while the remote hosts are lightweight cloud images
that only have a basic set of locales available (C.UTF-8, en_US.UTF-8,
etc.).

This mismatch then apparently causes the shell environment to
fallback to a **non-UTF-8 C locale**, wreaking all sorts of havoc,
especially with with log monitoring in Lorax/LMC, as the logs
*do* contain Unicode and make the log monitor crash due to Unicode
encoding errors.

So to avoid this nasty edge case, always set a well-known locale
to the LANG variable before starting remote jobs with Parallel.