Add pre-installation snapshot test
Before the installation this test will wipe existing things from the
disk (if there are any). Create a partition, make filesystem and store
a file. All this is done in the %pre section.

During the installation this partition is re-used for / so it is
formatted and all of the original content is stored only in the
snapshot.

After the installation snapshot is activated, mounted and tested.