Don't use inst.noninteractive by default for now
We are still hitting user-interaction-required race conditions
in GUI if inst.noninteractive is used for kickstart tests,
mostly for those with extensive %packages section.

While this is indeed an indication of a real bug, it also
prevents us from getting any results from package based tests.

So lets drop this from the default boo options for now
and put it back only once the race condition is fixed.