Check required packages are installed for firewall and team devices
The firewalld package should be installed for all installations that
contain the firewall command in input kickstart in any form.

The teamd package should be installed if any team devices have been
detected.

Both packages are needed to properly configure the target system
and these tests should make it easier to spot any issues related to
these packages either not being requested or being requested but not
installed for some reason.