In worker deployment task update packages first
In worker deployment task update packages first,
before installing groups and packages.

This way the system would be up to date, including DNF
and possibly avoiding some obscure but still nasty issues
with file conflicts that can happen when installing individual
groups or packages to a system that is out of date.