Run SElinux context test only after fixing contexts
The fixing of contexts happens in a custom %post script that is appended to
the kickstart read in. A %post script as used in this test previously -
directly in the kickstart file - was read before the fixing script, and
thus reported useless things.

Now the ks test includes the fixing script to ensure it is executed before
the actual test code.