Drop mount/root requirement from VirtualInstall
Use `blkid` to determine the label of the given ISO file instead of
pylorax. This eliminates the entire ISO mounting, `IsoDev` class, loop
device setup, and privilege requirement.

We also don't need the `.stage2` flag, as probe_boot_iso.sh *only*
accepts ISOs which have a stage2. (If that weren't so, we could make
probe_boot_iso.sh output that fact as a variable, and pass it to
VirtualInstall).