Fix runner container and make it unprivileged
The previous container with systemd libvirtd was not working any more,
as it tried to talk to the system D-Bus (and possibly other reasons).

Convert the container to run the tests as unprivileged `ksuser` instead,
which makes it palatable to safely run locally, as well as run on public
clouds.

With that, we don't need to start libvirt explicitly any more. The user
libvirtd starts automatically (via the library).