Don't copy boot.iso into output directory
This is a waste of space and time, symlinking it is enough.

It also does not matter when using the container, as only *.log files
are copied back to the host.

This was already applied earlier in commit b94e277562f8, but was
reverted in commit f6c89cdb1b48f as it broke in the old infra.