Add Fedora fast-path to probe_boot_iso.sh
Unpacking the three file systems is quite expensive and takes some 30s.
Avoid that for Fedora images which have a `./discinfo` file with the
release number in boot.iso.

This slightly changes the behaviour for Rawhide: /etc/os-release has
VERSION_ID=34 (at the time of writing), but ./discinfo has "Rawhide". As
`KSTEST_OS_VERSION` is not being used anywhere right now, this doesn't
break anything, and it may actually be useful for tests to know that
they are running on rawhide (as e.g. GPG key management is slightly
different there).