Run PR-affected tests in Travis
Travis now offers /dev/kvm, and its machines are powerful enough to run
at least a handful of tests. This is useful for PRs that change a few
tests only, to self-validate them.

Use the containers/runner/ environment, as some of the dependencies are
not available on Ubuntu (like lorax).

We can't run all tests on Travis, it would take too long. But at least
for gating PRs that add/change tests, it is enough to only run these.

If a PR does not change any tests, or does any change to the runner
logic (scripts/ or containers/), add a few representative tests to prove
that nothing broke too badly.