Add runner container launch script
This significantly simplifies the necessary setup steps to just one
command. From a fresh checkout you can run

    containers/runner/launch keyboard [... other tests]

It has sensible defaults, and works both as user and through sudo (for
user podman or and system podman/docker containers).

In "Configuration of the test" section, explain how to get an
interactive shell, as that's the most common "I need more control"
scenario.