Move kstests-in-cloud.sh into scripts/
This avoids run_kickstart_tests.sh's find_tests() to consider it as
test, and avoids this error:

```
make: Entering directory '/opt/kstest/kickstart-tests/scripts'
make: *** No rule to make target 'COMMAND'.  Stop.
```

his is a bit difficult to avoid, as we do want to keep it
executable, and find_tests() can't be changed to look for `*.ks.in`
instead as that would break setting `$KICKSTART_NAME` in a test (even
though none currently use it)

We could also rename it to drop the `.sh` suffix, but this seems to fit
better in the scripts/ directory anyway.