Support testing local updates.img files
If the run_kickstart_tests.sh `-u` (updates image) argument is a local
file, then start a simple python web server and let the QEMU guest use
the SLIRP IP of the host.

To support this via the runner container launch script, if the --updates
argument is a local file, then bind-mount it into the container.

This makes the validation of local anaconda changes really easy for a
developer or CI.