Pull in libvirt pool/volume TOCTOU fix
This often breaks our tests with non-trivial parallelism. Grab the fix
from GitHub as a patch, backport it to the older 2.2.1 version in Fedora
32, and hot-apply it in the container.

This will (and is supposed to) fail once a new libvirt version lands in
Fedora 32 with the fix. We will notice that in a failed
container-autoupdate.yml workflow, when that happens we can drop this
hack.