Support KSTEST_* substitution in fragments
Avoid duplicating the default Fedora rawhide repositories in the
default.ks fragment and scripts/default.sh. This makes it much easier to
set a local mirror in a test environment.

To implement this, run the %ksappend substitution before the
`@KSTEST_*@` sedding, so that the sed will catch the substitutions in
the fragments.

To make that work nicely, simplify apply-ksappend.py to only process one
file at a time (specified as positional command line argument, as it's
not really optional and thus should not be an option), avoid all the
file copying, and write the result to stdout, so that it can be
pipelined efficiently.

Also make any errors in apply-ksappend.py fatal. We don't want to hide
substitution errors or missing files in our tests, let's just fix them
instead.

Replace the ArgumentParser class in apply-ksappend.py with a simple
function. The class did not add any new functionality, and was just
unnecessary glue.