Expose runner container's libvirt and VNC
Start the user libvirt instance manually with `--listen`, so that it
listens to the TCP port. Disable TLS mode, as that requires setting up a
CA/certs; and all of this happens locally only anyway, so we don't need
authentication really.

Show a libvirt connection string to the user, so that it's easy to
control the container's libvirt on the host with `virsh -c`, or connect
to the running VMs with `virt-viewer -c` and visually follow the test
progress.

For now this only fully works with with system podman instances as they
use proper bridges.

For user podman instances we expose the port to the host, so that you
can find it in "podman ps". But this only works for controlling libvirt
itself, not for VNC.

But running as system instance is nicer anyway due to being able to use
the squid proxy, so that's not a priority for now.