Unify logging into kstest.log
Log everything from the controller into a single kstest.log file with a
consistent format. Due to using both our own logger and the pylorax
logging in parallel, the output was split between kstest-logger.log,
livemedia.log, and program.log; the first and third were lacking
timestamps, the first and second had a large overlap of logging the same
messages, and as a human it is hard to correctly interleave the messages
in your mind.

Also close the log after a test is done: Introduce a `close_logger()`
function and call it after Runner.run_test() is done. For symmetry, move
the opening of the log to that method as well. This isn't necessary for
te current run_one_test.py, but a prerequisite for introducing automatic
test retries.