Skip rhel8 tests that get broken by NetworkManager auto-connect regression
See <https://bugzilla.redhat.com/show_bug.cgi?id=1903061>. This breaks
many networking tests.

Mark these tests as `TESTTYPE=rhbz1903061`, and skip them from the
rhel8 scenario, so that it remains passing and we can thus spot other
regressions. This also provides an easy way to run just these tests to
validate a fix.

This is a generic pattern which we can use for other OS regressions as
well.