Add nightly testing of Anaconda RHEL COPR build
Use lorax to build a boot.iso from current nightly RHEL 8 and our
nightly anaconda COPR, and run the "coverage" type tests on it.

This all happens in one workflow, as otherwise we would need to find a
place to put the boot.iso. We can't put it into a GitHub artifact as
that's world-readable, and we can't just leave it on the host machine as
the build and the run workflow may happen on different machines. So
start like that, it can be refined later if we want to use some internal
network storage. For local debugging, the current image can be
downloaded from the github user's home directory.

Fixes #427