Fix issue when containers does not have access to /dev/kvm
This is happening when podman is started in rootless mode. In that case /dev/kvm
socket has wrong SELinux label because it just bind-mounted inside of the
container. For more info see here:

https://bugzilla.redhat.com/show_bug.cgi?id=1901462

To solve this issue we need to add '--security-opt label=disabled' which will
disable container SELinux separation. This will be used only if we are runnig
podman in a rootless mode (use 'sudo podman' to avoid that).