Fix squid cleanup if container is stopped
It may happen that the squid container is not running any more, but
still exists in the "exited" state. In that case, stop() will not do
anything, but removing the cache volume fails because it's still being
used by the stopped container.

Force the removal of the container in clean() to avoid that.