Fix network interface naming in daily-boot-iso-rhel8 workflow
For some strange reason, the path vs. slot based network interface
naming depends on the iso's CD volume ID. Move to
"RHEL-8-0-0-BaseOS-x86_64" to match the label structure of the official
iso while still avoiding the hardcoding of "8.4" (it seems it is enough
to be a digit, but which particular one does not matter).

Revert the previous commit 918215f6ce3 hack, and go back to testing both
the official and our daily image with the same defaults-rhel8.sh.

Fixes #448