Additional RPM repo library functions
Add a couple library functions that make it easy to include additional
RPM repositories to a kickstart test.

By using these functions in it's sh script a kickstart tests can easily
add arbitrary RPM packages from the data folder to the installation
transaction. This can be useful to test a couple package scratchbuilds
without having to manipulate the full installation repos to do this.

The mechanism works by checking if the "additional_repo" folder exists
in the data folder. If it does a localhost HTTP server is started and
it's address is written to the addrepo_url file in the tests tempdir.

NOTE: We expect the additional_repo folder to just contain a bunch of
RPM files. The library function will make sure to run createrepo as
needed on it.

There is then a second library function that checks if the addrepo_url
file exists and adds appropriate inst.addrepo invocation to the passed
boot option script.

But functions still consider the additional repo to be optional,
so if no additional_repo exists in the data folder, they will be
effectively a no-op and the kickstart test using them will continue to
execute as usual, of course without the additional packages being added.