Use additional RPM repos in Initial Setup reboot tests
Use the library functions that enable access to additional RPM
repos in the Initial Setup reboot tests.

These tests aim is to verify that latest Anaconda changes have not
broken Initial Setup. To do this they need an Anaconda scratchbuild with
the latest changes (usually still in PR form) to be injected to the
package installation transaction as otherwise the stable released
package from the main repo would be installed and no breakage could
be uncovered.