Install the modular repository
Install the modular repository with a higher priority on the target system,
so DNF can use it in %post scripts. The repository that is installed by
a package might not be available and cause a failure of the test.

The original repository
that is installed by a package, might not be available.