Make it possible to set per-test timeout
Some kickstart tests might need more time to run to the end
than others in the same batch. In other cases we might want the opposite
- if the test does not end in a few minutes, something is wrong and
waiting for an hour will not help.

To do this a new function called timeout() has been added, that can be
defined in a kickstart test utility script, to set a custom test
timeout.

The timeout() function is expected to return the number of minutes
as an integer in string form. If a test does not define the timeout()
method, the current default value of 60 minutes will be used.