Fix module installation tests 1-4
The unstoppable march of postgress yet again resulted in streams
that were once new and shiny being dead and forgotten and replaced
by streams that are new and shiny.

So fixup the module-1 to 4 tests with new and shiny nodejs and
postgressql stream names, fixing the tests once and for all (of November)!

Fixes #575