Add log2json script and pass scenario name to it
This adds the ability to pass the scenario name into run-kstest so that
the log summary can group the test results based on which scenario is
being run.

It adds the log2json script, and calls it on the combined kstest.log
file in order to generate a JSON summary of all of the tests.

The github workflow has also been modified to archive this new
kstest.log.json file.