weekly-summary: Add more information about the tests
This adds a count of the 'flaky' tests, ones which fail at first but
eventually succeed. This now counts as a successful test, but is also
listed separately.

This adds an option to include detailed info about the flaky tests if
you pass `--flaky-details` to the script.

This now handles expired artifacts -- GitHub is not returning an error
if you try to retrieve an expired artifact, so it needs to handle trying
to unzip a non-zipfile (it's a json blob).

Remove unused functions, and split up some of the logic into functions.
Also improves documentation of the data structures created for
generating the reports and stats.