weekly-summary: Make an archive of failed/flake logs
If --archive-logs /var/tmp/kstest-log-archive is passed the summary will
create a directory tree matching the summary text hierarchy that
contains the raw logs for the failed and flaky tests. This requires
using the original raw archives, not the daily json summary archives.

The layout for the directory looks like this:

    MM-DD-YYYY / scenario / failed|flakes / TESTNAME / [1...] /