Save the list of  expected tests to kstest-list
There are situations where a test is not started and doesn't generate a
log directory or file. The only way to know that test is missing is if
there is a list of all the expected tests for the current scenario. This
saves that list as kstest-list and adds it to the archive for later use
by log2json.