weekly-summary: Add support for pages and rate limits
The API query only returns 100 results at a time. All of the artifacts
may not be included in the first page.

This adds support for requesting all the pages and retrieving artifacts
within the date range and desired names from each page.

This also adds a check for the GitHub rate limits. If is is hit, and the
reset time is less than 90s in the future it will wait and try again. If
this fails 3 times it will then exit with an error.