Collect virt-install log when running a test.
The log is useful when the test fails with "Problem starting virtual
install".

The log will not be collected from remote runners. The remote runners
functionality (TEST_REMOTES enviroment variable) is not used since a few
years ago as the distribution of the tests to runners would be done on
higher level now.