Introduce a pylint check
Also switch the test to run on our container for gating purposes and
also for having the python modules installed, so that pylint won't
complain.

Stop specifying the timeout, both ruff and pylint are fast enough for
the timeout to not be relevant.

Lastly remove the ENV variables from the static-code-tests workflow, these are also
not relevant.