Add support for RDP based remote desktop access
Test the new RDP based remote desktop support that gets in together with
the Wayland support patches.

NOTE: Right now we are testing that RDP server is running and and
default system target is set correctly. But we are not yet testing
if it is actually possible to connect to the RDP session with a client.