Add test for UEFI bootupd with leavebootorder
The rpm-ostree-container installations will by default install an UEFI
entry which did not happened before. However, if leavebootorder is set
it shouldn't happen.

On RHEL-9 and RHEL-10 the leavebootorder will be ignored because it's
not yet implemented, however, the final behavior is the same so let's
leave it enabled also on these platforms.