Fix error when using different boot.iso name
We have a code for specifying a boot.iso name for a long time but thanks
to this line it didn't worked. Reason is this code is working in tmpfs
space which has only a symlink of the original boot.iso with the static
boot.iso name.

This bug is here for a long time but we never used a different boot.iso
name. Let's fix that and use it to our benefit.